requirements.txt
flask
firebase-admin
python-dotenv
cloudinary
flask-login
gunicorn